requirements.txt
# Core web framework
streamlit>=1.28.0

# HTTP requests and web frameworks
requests>=2.31.0
flask>=3.0.0

# Environment management
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.26.0,<2

# Vector database and embeddings
faiss-cpu>=1.7.4
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-aws>=0.2.0
langchain-core>=0.3.0

# Azure services (for embeddings only)
azure-identity>=1.15.0
azure-core>=1.30.0

# AWS services
boto3>=1.34.0

# Text processing and AI
tiktoken>=0.5.0

# Machine Learning and NLP
sentence-transformers>=2.2.0
torch>=2.0.0
rank-bm25>=0.2.2

# File processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Image processing
Pillow>=10.0.0

# Document processing
docling>=2.0.0
docling-core>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21

# FastAPI and server components
fastapi>=0.104.0
uvicorn>=0.24.0

# Natural language processing
nltk>=3.8.0

# Utilities
tqdm>=4.66.0
typing-extensions>=4.8.0

# Data visualization
altair>=5.0.0

# Additional utilities that might be imported
beautifulsoup4>=4.12.0
faiss-cpu>=1.7.4